refactor(file): hooks 내에 pages 추가
